requirements.txt
streamlit
pypdf
langchain
langchain-community
langchain-text-splitters
faiss-cpu
google-genai
python-dotenv
sentence-transformers